website/streamlit-prototype/requirements.txt
streamlit
pandas
numpy
plotly
requests
fastapi
uvicorn
sqlalchemy
pydantic-settings
loguru
